+++ packages/database-devtools/vitest/mocks/material-community-icons.ts
@@ -1,0 +1,3 @@
+const MaterialCommunityIcons = 'MaterialCommunityIcons';
+
+export default MaterialCommunityIcons;